refactor(tv-next): reuse common types
