+++ packages/mirror-tv-next/types/common.ts
@@ -3,6 +3,13 @@ export type FlashNews = {
   name: string
 }
 
+export type HeroImage = {
+  urlDesktopSized: string
+  urlTabletSized: string
+  urlMobileSized: string
+  urlOriginal: string
+}
+
 export type Schedule = {
   'Channel ID': string
   Year: string
